added base redis queue that circular queue inherits from
